resume-checker/requirements.txt
streamlit
pdfminer.six
spacy
scikit-learn
sentence-transformers
numpy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl